Move array_ops tests out of ctest_array and into ctest_array_ops.
